requirements.txt
altgraph==0.17
aniso8601==7.0.0
antiorm==1.2.1
appdirs==1.4.4
beautifulsoup4==4.9.0
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.1
dominate==2.6.0
ez-setup==0.9
filelock==3.0.12
flake8==3.9.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-talisman==0.7.0
future==0.18.2
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
liqpay-python==1.0
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.5
pandas==1.2.0
paypalhttp==1.0.0
pefile==2019.4.18
Pillow==7.2.0
pipenv==2020.11.15
promise==2.3
psycopg2==2.8.5
py4j==0.10.9
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pymongo==3.11.0
PyMySQL==0.10.0
pyOpenSSL==19.1.0
pyspark==3.0.1
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.23.0
Rx==1.6.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0
SQLAlchemy==1.3.20
transliterate==1.10.2
urllib3==1.25.9
vcrpy==4.1.1
vcrpy-unittest==0.1.7
virtualenv==20.1.0
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.2.8
yarl==1.6.3
